star-257 create enum helper
