+++ packages/Card/src/Card.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import * as sc from "./Card.styles";
 import Footer from "./components/Footer";
 import Title from "./components/Title";
