+++ packages/Checkbox/stories/examples/Showcase.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { boolean, select, text } from "@storybook/addon-knobs";
 import { Rule, Tagline } from "storybook/assets/styles/common.styles";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import Heading from "@paprika/heading";
 import { CheckboxStory } from "../Checkbox.stories.styles";
 import Checkbox from "../../src/Checkbox";
