+++ packages/Counter/stories/examples/Showcase.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { boolean, number, select } from "@storybook/addon-knobs";
 import { Story, Rule, Tagline } from "storybook/assets/styles/common.styles";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import Heading from "@paprika/heading";
 import Counter from "../../src/Counter";
