+++ packages/DatePicker/src/components/DateInputPropsCollector/DateInputPropsCollector.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 
 const propTypes = {
   /** a11yText on the input. */
