+++ packages/DropdownMenu/stories/examples/Showcase.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Rule, Tagline } from "storybook/assets/styles/common.styles";
 import { select, text } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
-import { AlignTypes } from "@paprika/helpers/lib/customPropTypes";
+import { AlignTypes } from "@paprika/helpers/lib/enums";
 import L10n from "@paprika/l10n";
 import Heading from "@paprika/heading";
 import Confirmation from "@paprika/confirmation";
