+++ packages/FormElement/stories/examples/Showcase.js
@@ -3,7 +3,7 @@ import { boolean, select, text } from "@storybook/addon-knobs";
 import Heading from "@paprika/heading";
 import Input from "@paprika/input";
 import { Rule, Tagline } from "storybook/assets/styles/common.styles";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import { FormElementStory } from "../FormElement.stories.styles";
 import FormElement from "../../src";
