+++ packages/Input/src/Input.styles.js
@@ -1,7 +1,7 @@
 import { css } from "styled-components";
 import tokens from "@paprika/tokens";
 import stylers from "@paprika/stylers";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 
 const iconStyles = {
   [ShirtSizes.SMALL]: `
