+++ packages/ListBox/src/ListBox.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import useI18n from "@paprika/l10n/lib/useI18n";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import Box from "./components/Box";
 import Content from "./components/Content";
 import List from "./components/List";
