+++ packages/Modal/src/Modal.styles.js
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
 import tokens from "@paprika/tokens";
 import stylers from "@paprika/stylers";
-import { ShirtSizes } from "@paprika/helpers/lib/customPropTypes";
+import { ShirtSizes } from "@paprika/helpers/lib/enums";
 import OriginalHeader from "./components/Header";
 import OriginalContent from "./components/Content";
 import OriginalFooter from "./components/Footer";
